refactor: reorganize code - clean architecture
