+++ Controllers/ProjectsController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.RateLimiting;
-using ProjectManager_01.WebAPI.Data;
+using ProjectManager_01.WebAPI.Domain.Models;
 
 namespace ProjectManager_01.WebAPI.Controllers;
