+++ Controllers/UserRolesController.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.RateLimiting;
-using ProjectManager_01.WebAPI.Data;
-using ProjectManager_01.WebAPI.Enums;
+using ProjectManager_01.WebAPI.Domain.Models;
+using ProjectManager_01.WebAPI.Domain.Enums;
 
 namespace ProjectManager_01.WebAPI.Controllers;
