+++ Hubs/TicketsHub.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.AspNetCore.SignalR;
-using ProjectManager_01.WebAPI.Data;
+using ProjectManager_01.WebAPI.Domain.Models;
 
 namespace ProjectManager_01.WebAPI.Hubs;
