+++ ProjectManager-01.WebAPI-Domain/Models/Permission.cs
@@ -1,4 +1,4 @@
-﻿namespace ProjectManager_01.WebAPI.Data;
+﻿namespace ProjectManager_01.WebAPI.Domain.Models;
 
 public sealed class Permission
 {
