+++ ProjectManager-01.WebAPI-Domain/Models/UserRole.cs
@@ -1,6 +1,6 @@
-﻿using ProjectManager_01.WebAPI.Enums;
+﻿using ProjectManager_01.WebAPI.Domain.Enums;
 
-namespace ProjectManager_01.WebAPI.Data;
+namespace ProjectManager_01.WebAPI.Domain.Models;
 
 public sealed class UserRole
 {
